Split heartbeat prompt smoke fixtures
